flurry cast moved from brain freeze to flurry module
