+++ src/analysis/retail/mage/frost/Guide.tsx
@@ -6,6 +6,7 @@ import { formatPercentage } from 'common/format';
 import ActiveTimeGraph from 'parser/ui/ActiveTimeGraph';
 import { SpellLink } from 'interface';
 import TALENTS from 'common/TALENTS/mage';
+import PreparationSection from 'interface/guide/components/Preparation/PreparationSection';
 
 export const GUIDE_CORE_EXPLANATION_PERCENT = 50;
 
@@ -49,6 +50,7 @@ export default function Guide({ modules, events, info }: GuideProps<typeof Comba
     <>
       <Section title="Core">
         {modules.wintersChill.guideSubsection}
+        {modules.flurry.guideSubsection}
         {alwaysBeCastingSubsection}
       </Section>
       <Section title="Cooldowns">
@@ -56,7 +58,11 @@ export default function Guide({ modules, events, info }: GuideProps<typeof Comba
         {info.combatant.hasTalent(TALENTS.RAY_OF_FROST_TALENT) &&
           modules.rayOfFrost.guideSubsection}
         {info.combatant.hasTalent(TALENTS.COMET_STORM_TALENT) && modules.cometStorm.guideSubsection}
+        {info.combatant.hasTalent(TALENTS.FROZEN_ORB_TALENT) && modules.frozenOrb.guideSubsection}
+        {info.combatant.hasTalent(TALENTS.SHIFTING_POWER_TALENT) &&
+          modules.shiftingPowerFrost.guideSubsection}
       </Section>
+      <PreparationSection />
     </>
   );
 }
